Hardcaml Zprize Submission

This documentation describes two projects undertaken as part of the Zprize competition in 2022.

Multi-scalar multiplication

Multiply 226 BLS12-377 affine points by 253 bit scalars and add the result.

Pippenger implements the controller logic to sequence the MSM computation.

Elliptic_curve_lib, Field_ops_lib and Twisted_edwards_lib implement the core elliptic curve arithmetic operations.

Msm_pippenger implements the complete MSM computation using the Pippenger bucket method.

Inverse NTT

Perform 224 inverse number theoretic transform over a 64 bit finite field.

Hardcaml_ntt implements the core components of for the NTT transform.

Zprize_ntt implements the top level design and dicsusses results.

Support libraries

Hardcaml_axi

Ark_bls12_377_g1

Twisted_edwards_lib

Field_ops_lib - Library for operations in a Galois fields (eg: large-bit-width multiplication)